Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    float a,b,c;
+    int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
     float f;
-    f = (a + b + c)/3;
+    f = (a + b + c)/3.0;
     printf("Average: %.2f",f);
     return 0;
 }
